requirements.txt
Django==4.2.7
google-genai==0.3.0
python-dotenv==1.0.0
pyautogui==0.9.54
speechrecognition==3.10.0
pyttsx3==2.90
opencv-python==4.8.0.76
numpy==1.26.4